==> _Code/4_gendered_words_list.R <==
# load necessary libraries
library(webdriver)
require(webdriver)
install_phantomjs()
require(rvest)
library(stringr)
library(splitstackshape)

# in this script, the list with the gendered words is generated by scraping
# the websites "geschicktgendern.de" and "gendern.de"

# start phantomjs
pjs_instance <- run_phantomjs()
pjs_session <- Session$new(port = pjs_instance$port)

#---------- get gendered words for geschicktgendern.de
# set URL
url <- "https://geschicktgendern.de/"

# load URL to phantomJS session
pjs_session$go(url)
# retrieve the rendered source code of the page
rendered_source <- pjs_session$getSource()
# parse the dynamically rendered source code
html_document <- read_html(rendered_source)
# set text path
text_path <- "//tbody[contains(@class, 'row-hover')]//td"

title_text <- html_document %>%
  html_nodes(xpath = text_path) %>%
  html_text(trim = T)

# remove words that are not gendered
gendered_words1 <- title_text[c(FALSE,TRUE)]
# create dataframe
gendered_words1 <- data.frame(gendered_words1)
# rename column for rbind later
names(gendered_words1)[1] <- "gendered_words"

#---------- get gendered words for gendern.de

# Send a GET request to the URL
url <- "https://www.gendern.de/"
page <- read_html(url)



# Create an empty data frame to store the table rows
gendered_words2 <- data.frame()

# Loop through each page
while (TRUE) {
  
  # Extract the table body from the current page
  table <- page %>% html_nodes("table") %>% html_table(fill = TRUE)
  gendered_words2 <- rbind(gendered_words2, table)
  
  # Look for the button that loads the next page
  next_button <- page %>% html_nodes("button:contains('Weiter')")
  
  # Exit the loop if there's no more pages to load
  if (length(next_button) == 0) {
    break
  }
  
  # Extract the attributes of the next page URL and construct the URL
  next_url <- paste0(url, html_attr(next_button, "onclick"))
  next_url <- gsub(".*\\('", "", next_url)
  next_url <- gsub("'.*", "", next_url)
  
  # Send a GET request to the next page URL
  page <- read_html(next_url)
}

# choose only relevant column with gendered words
gendered_words2 <- data.frame(gendered_words2[,2])
# rename column for rbind later
names(gendered_words2)[1] <- "gendered_words"

#---------- combine both dfs and eliminate duplicates

# combine both dfs
gendered_words <- rbind(gendered_words1, gendered_words2) #8454 rows
# remove duplicates
gendered_words <- gendered_words[!duplicated(gendered_words),] #6245 rows
gendered_words <- data.frame(gendered_words)
# order alphabetically
gendered_words <- gendered_words[order(gendered_words$gendered_words),]
gendered_words <- data.frame(gendered_words)

#---------- clean the df of gendered words
# separate strings at ";"
gendered_words_splitted <- cSplit(gendered_words, "gendered_words", sep = ";", direction = "long")
# remove rows that contain "[" or "]", because they don't contain genderneutral terms
gendered_words_splitted <- gendered_words_splitted[!grepl("\\[|\\]", gendered_words_splitted$gendered_words),]
# remove "..."
gendered_words_splitted <- gendered_words_splitted[!grepl("\\...", gendered_words_splitted$gendered_words),]
# remove rows that contain more than one string
gendered_words_splitted <- gendered_words_splitted[!apply(gendered_words_splitted, 1, function(x) any(lengths(strsplit(as.character(x), " ")) > 1)),]
# remove rows where any cell starts with a lowercase character
gendered_words_splitted <- gendered_words_splitted[!apply(gendered_words_splitted, 1, function(x) any(grepl("^[[:lower:]]", as.character(x)))),]
# remove rows which contain "*", "_" or "/" in between the word
gendered_words_splitted <- gendered_words_splitted[!grepl("^[^*/_]*[*_/][^*/_]*$", gendered_words_splitted$gendered_words), ]
# remove duplicates
gendered_words_splitted <- gendered_words_splitted[!duplicated(gendered_words_splitted),]
# remove empty rows
gendered_words_splitted <- gendered_words_splitted[!apply(gendered_words_splitted == "", 1, all),] #1566 words

# export csv 
write.csv(gendered_words_splitted, "Genderliste\\gendered_words_splitted.csv")
